requirements.txt
numpy==1.23.3
pygame==2.1.1
PyOpenGL==3.1.5